# backend/openapi.yaml
openapi: "3.0.3"
info:
  title: AgentID 2.0 API
  version: "2.0.0"
  description: |
    AgentID 2.0 — Universal Non-Human Identity (NHI) platform.
    Provides agent registration, reputation scoring, W3C DID/VC credential export,
    A2A authentication tokens, organization management, policy enforcement,
    webhook integrations, and embeddable trust badges/widgets.

    All routes are available at both the root path and under the `/v1/` prefix.
    Root-mounted routes are deprecated; prefer `/v1/` for new integrations.
  contact:
    name: AgentID Team
    url: https://agentidapp.com
  license:
    name: MIT
servers:
  - url: https://api.agentidapp.com
    description: Production
  - url: https://api.agentidapp.com/v1
    description: Production (versioned)
  - url: http://localhost:3002
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /verify-token:
    post:
      tags: [Agents]
      operationId: verifyA2AToken
      summary: Verify an A2A authentication token (public)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        "200":
          description: Token verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid: { type: boolean }
                  agentId: { type: string }
                  claims: { type: object }
...
  /credentials/verify:
    post:
      tags: [Credentials]
      operationId: verifyCredential
      summary: Verify a W3C Verifiable Credential (F-02)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [credential]
              properties:
                credential: { $ref: "#/components/schemas/VerifiableCredential" }
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid: { type: boolean }
                  checks: { type: array, items: { type: string } }
                  errors: { type: array, items: { type: string } }

  # ── Audit ─────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  schemas:
...
    VerifiableCredential:
      type: object
      properties:
        "@context": { type: array, items: { type: string } }
        type: { type: array, items: { type: string } }
        issuer: { type: string }
        issuanceDate: { type: string, format: date-time }
        credentialSubject:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
            reputationScore: { type: number }
        proof:
          type: object
          properties:
            type: { type: string }
            created: { type: string, format: date-time }
            verificationMethod: { type: string }
            proofPurpose: { type: string }
            proofValue: { type: string }
...